+++ scrapli/exceptions.py
@@ -43,3 +43,7 @@ class TransportPluginError(ScrapliException):
 
 class ConnectionNotOpened(ScrapliException):
     """Exception for failures due to connection not being open"""
+
+
+class UnsupportedPlatform(ScrapliException):
+    """Exception for unsupported platform; i.e. using system transport on windows"""